Extract mutations.py from legacy.py
